Changed horizon that it has started display time frames, but not correct

+++ src/github.com/stellar/horizon/resource/signers_access.go
@@ -10,7 +10,7 @@ import (
 func (this *SignersAccess) Populate(ctx context.Context, row core.SignersAccess) {
 	this.AccessGiverID = row.Accessgiverid
 	this.AccessTakerID = row.Accesstakerid
-	this.TimeFrames = row.Timeframes
+	this.TimeFrames = int64(row.Timeframes)
 
 	lb := hal.LinkBuilder{httpx.BaseURL(ctx)}
 	this.Links.Self = lb.Linkf("/signersaccesses/%d", row.Accesstakerid, row.Accessgiverid, row.Timeframes)
